+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 1.0.5
+  - Add pt-BR language pack, by @thugking from Rchan.
   - Fix bypass not auto closing and submitting on non english locale.
   - Actually fix Tor hidden service country name not displaying correctly.
   - Make the error handling for "possible corrupt or unsupported file" catch more bad files and processing errors.